+++ src/main/scala/org/hammerlab/magic/args4j/StringOptionHandler.scala
@@ -3,6 +3,9 @@ package org.hammerlab.magic.args4j
 import org.kohsuke.args4j.spi.{OptionHandler, Parameters, Setter}
 import org.kohsuke.args4j.{CmdLineParser, OptionDef, OptionHandlerRegistry}
 
+/**
+ * Args4J option handler that populates an [[Option[String]]].
+ */
 class StringOptionHandler(parser: CmdLineParser, option: OptionDef, setter: Setter[Option[String]])
   extends OptionHandler[Option[String]](parser, option, setter) {
   override def getDefaultMetaVariable: String = "path"
@@ -14,7 +17,5 @@ class StringOptionHandler(parser: CmdLineParser, option: OptionDef, setter: Sett
 }
 
 object StringOptionHandler {
-  println("registering StringOptionHandler")
   OptionHandlerRegistry.getRegistry.registerHandler(classOf[Option[String]], classOf[StringOptionHandler])
-  println("registered")
 }
