+++ src/main/scala/org/hammerlab/magic/iterator/LinesIterator.scala
@@ -2,6 +2,9 @@ package org.hammerlab.magic.iterator
 
 import java.io.{BufferedReader, InputStream, InputStreamReader}
 
+/**
+ * Wrap a [[BufferedReader]] and emit lines.
+ */
 class LinesIterator(br: BufferedReader) extends SimpleBufferedIterator[String] {
   override def _advance: Option[String] =
     if (br.ready)
