+++ src/main/scala/org/hammerlab/magic/iterator/RangeAccruingIterator.scala
@@ -1,5 +1,14 @@
 package org.hammerlab.magic.iterator
 
+/**
+ * Given an [[Iterator]] of [[Int]]s, collapse contiguous "ranges" of integers that are each 1 greater than their
+ * predecessor.
+ *
+ * For example, given an input [2, 3, 1, 4, 5, 6, 5, 6, 8], this would emit [[Range]]s (in half-open notation): [2, 4),
+ * [1, 2), [4, 7), [5, 7), [8, 9).
+ *
+ * See RangeAccruingIteratorTest for more examples.
+ */
 class RangeAccruingIterator(it: Iterator[Int]) extends Iterator[Range] {
 
   var anchor = -1
