+++ src/main/scala/org/hammerlab/magic/iterator/RunLengthIterator.scala
@@ -3,6 +3,12 @@ package org.hammerlab.magic.iterator
 import spire.math.Integral
 import spire.implicits._
 
+/**
+ * Run-length encode an input iterator, replacing contiguous runs of identical elements with pairs consisting of the
+ * first element in the run and the number of elements observed.
+ *
+ * See RunLengthIteratorTest for examples.
+ */
 class RunLengthIterator[K] private(it: BufferedIterator[K]) extends Iterator[(K, Int)] {
 
   override def hasNext: Boolean = it.hasNext
