+++ src/main/scala/org/hammerlab/magic/kryo/Registrar.scala
@@ -2,7 +2,8 @@ package org.hammerlab.magic.kryo
 
 import com.esotericsoftware.kryo.Kryo
 import org.apache.spark.serializer.KryoRegistrator
-import org.hammerlab.magic.rdd.{KeySamples, RDDStats, RunLengthRDD}
+import org.hammerlab.magic.rdd.groupby.KeySamples
+import org.hammerlab.magic.rdd.{RDDStats, RunLengthRDD}
 
 class Registrar extends KryoRegistrator {
   override def registerClasses(kryo: Kryo): Unit = {
