+++ src/main/scala/org/hammerlab/magic/rdd/CollectPartitionsRDD.scala
@@ -4,6 +4,10 @@ import org.apache.spark.rdd.RDD
 
 import scala.reflect.ClassTag
 
+/**
+ * Wrap an [[RDD]] and expose a `collectPartitions` method that is similar to [[RDD.collect]], but returns an [[Array]]
+ * of per-partition [[Array]]s.
+ */
 class CollectPartitionsRDD[T: ClassTag](@transient rdd: RDD[T]) extends Serializable {
   def collectPartitions(): Array[Array[T]] = rdd.mapPartitions(it => Iterator(it.toArray)).collect()
 }
