+++ src/main/scala/org/hammerlab/magic/rdd/ReduceByKeyRDD.scala
@@ -4,6 +4,9 @@ import org.apache.spark.rdd.RDD
 
 import scala.reflect.ClassTag
 
+/**
+ * Adds [[maxByKey]] and [[minByKey]] helpers to an [[RDD]].
+ */
 class ReduceByKeyRDD[K: ClassTag, V: ClassTag](rdd: RDD[(K, V)])(implicit ord: Ordering[V]) {
   def maxByKey(numPartitions: Int = rdd.getNumPartitions): RDD[(K, V)] = rdd.reduceByKey((a, b) => ord.max(a, b))
   def minByKey(numPartitions: Int = rdd.getNumPartitions): RDD[(K, V)] = rdd.reduceByKey((a, b) => ord.min(a, b))
