+++ src/main/scala/org/hammerlab/magic/rdd/SlidingRDD.scala
@@ -6,6 +6,10 @@ import org.hammerlab.magic.rdd.BorrowElemsRDD._
 
 import scala.reflect.ClassTag
 
+/**
+ * Helpers for mimicking Scala collections' "sliding" API on [[RDD]]s; iterates over successive N-element subsequences
+ * of the [[RDD]].
+ */
 class SlidingRDD[T: ClassTag](@transient rdd: RDD[T]) extends Serializable {
   def sliding2(fill: T): RDD[(T, T)] = sliding2(Some(fill))
   def sliding2(fillOpt: Option[T] = None): RDD[(T, T)] =
