+++ src/main/scala/org/hammerlab/magic/rdd/grid/Summable.scala
@@ -1,0 +1,9 @@
+package org.hammerlab.magic.rdd.grid
+
+trait Summable[+Companion <: SummableCompanion[_]] {
+  def +(o: Summable[_]): this.type
+}
+
+trait SummableCompanion[+S <: Summable[_]] {
+  def zero: S
+}
