+++ src/main/scala/org/hammerlab/magic/util/KeyOrdering.scala
@@ -1,5 +1,8 @@
 package org.hammerlab.magic.util
 
+/**
+ * Order [[Tuple2]]s by key.
+ */
 class KeyOrdering[T, U](ordering: Ordering[T]) extends Ordering[(T, U)] {
   override def compare(x: (T, U), y: (T, U)): Int = ordering.compare(x._1, y._1)
 }
